[stdlib] Split ObjC parts out of DictionaryTraps test
